+++ R/NCI60.R
@@ -1,18 +1,19 @@
-
 #' Load the NCI60 data from GitHub
-#' 
-#' @importFrom utils download.file
 #' @export
 #' @examples 
 #' 
+#' \dontrun{
 #' NCI60 <- etl_NCI60()
+#' }
 
 etl_NCI60 <- function() {
   NCI60 <- data.frame()
   dir <- tempdir()
   lcl <- file.path(dir, "NCI60.rda")
-  utils::download.file("https://github.com/beanumber/mdsr/blob/master/data-raw/NCI60.rda?raw=true", 
-                destfile = lcl)
+  downloader::download(
+    "https://github.com/beanumber/mdsr/blob/master/data-raw/NCI60.rda?raw=true", 
+    destfile = lcl
+  )
   load(lcl)
-  return(dplyr::as.tbl(NCI60))
+  return(tibble::as_tibble(NCI60))
 }