+++ R/scidb.R
@@ -1,18 +1,11 @@
 #' src_scidb
 #' @description Connect to the scidb server at Smith College.
-#' 
 #' @param dbname the name of the database to which you want to connect
 #' @param ... arguments passed to \code{\link[dplyr]{src_mysql}} or \code{\link[DBI]{dbConnect}}
-#' 
 #' @details This is a public, read-only account. Any abuse will be considered a 
 #' hostile act. 
-#' 
-#' @return For \code{\link{src_mysql}}, a \code{\link[dplyr]{src_mysql}} object
-#' 
-#' @seealso \code{\link[dplyr]{src_mysql}}
-#' 
-#' @importFrom dplyr src_mysql
-#' @importFrom RMySQL MySQL
+#' @return For \code{\link{src_scidb}}, a \code{\link[dplyr]{src_dbi}} object
+#' @seealso \code{\link[dbplyr]{src_dbi}}
 #' @export
 #' 
 #' @examples 
@@ -21,35 +14,29 @@
 #' dbAir
 
 src_scidb <- function(dbname, ...) {
-  dplyr::src_mysql(dbname, host = "scidb.smith.edu", 
-                   user = "mth292", password = "RememberPi")
+  dbplyr::src_dbi(dbConnect_scidb(dbname, ...))
 }
 
 #' @rdname src_scidb
 #' @export
-#'  
-#' @importFrom DBI dbConnect
-#' @importFrom RMySQL MySQL
-#' 
 #' @return For \code{\link{dbConnect_scidb}}, a \code{\link[DBI]{DBIConnection-class}} object
 #' 
 #' @seealso \code{\link[DBI]{DBIConnection-class}}
 #' @examples 
-#' 
 #' dbAir <- dbConnect_scidb("airlines")
 #' dbAir
 
 dbConnect_scidb <- function(dbname, ...) {
-  dbConnect(RMySQL::MySQL(), dbname = dbname, host = "scidb.smith.edu", 
-            user = "mth292", password = "RememberPi")
+  DBI::dbConnect(RMySQL::MySQL(), dbname = dbname, 
+                 host = "mdsr.cdc7tgkkqd0n.us-east-1.rds.amazonaws.com", 
+                 user = "mdsr_public", password = "ImhsmflMDSwR")
 }
 
 #' @rdname src_scidb
 #' @export
 #' 
-#' @return For \code{\link{mysql_scidb}}, a charcter vector of length 1 to be used
+#' @return For \code{\link{mysql_scidb}}, a character vector of length 1 to be used
 #' as an \code{engine.ops} argument, or on the command line. 
-#' 
 #' @seealso \code{\link[knitr]{opts_chunk}}
 #' @examples 
 #' 
@@ -58,5 +45,5 @@ dbConnect_scidb <- function(dbname, ...) {
 #' }
 
 mysql_scidb <- function(dbname, ...) {
-  paste("-h scidb.smith.edu -u mth292 -pRememberPi -t", dbname)
+  paste("-h mdsr.cdc7tgkkqd0n.us-east-1.rds.amazonaws.com -u mdsr_public -pImhsmflMDSwR -t", dbname)
 }