+++ R/themes.R
@@ -3,15 +3,15 @@
 #' @description Graphical themes used in MDSR book
 #' 
 #' @inheritParams ggplot2::theme_grey
-#' @importFrom ggplot2 theme theme_grey element_text element_line element_rect rel %+replace%
-#' @import mosaic
+#' @importFrom ggplot2 %+replace%
 #' @export
 #' @examples 
-#' 
-#' p <- ggplot(mtcars, aes(x = hp, y = mpg, color = factor(cyl))) + 
-#'   geom_point() + facet_wrap(~ am) + geom_smooth()
-#' p + theme_grey()
-#' p + theme_mdsr()
+#' if (require(ggplot2)) {
+#'   p <- ggplot(mtcars, aes(x = hp, y = mpg, color = factor(cyl))) + 
+#'     geom_point() + facet_wrap(~ am) + geom_smooth()
+#'   p + theme_grey()
+#'   p + theme_mdsr()
+#'  }
 
 theme_mdsr <- function(base_size = 12, base_family = "Bookman") {
   m <- mosaic::theme_map()
@@ -26,4 +26,25 @@ theme_mdsr <- function(base_size = 12, base_family = "Bookman") {
       panel.grid.minor = ggplot2::element_line(colour = "grey92", size = 0.5),
       strip.background = ggplot2::element_rect(fill = "grey80", colour = "grey50", size = 0.2)
     )
+}
+
+#' Custom skimmer
+#' @export
+#' @importFrom skimr skim
+#' @inheritParams skimr::skim
+#' @examples
+#' skim(faithful)
+
+skim <- function(data, ...) {
+  my_skim <- skimr::skim_with(
+    base = skimr::sfl(
+      n = length,
+      missing = skimr::n_missing
+    ), 
+    numeric = skimr::sfl(
+      hist = NULL
+    )
+  )
+  my_skim(data, ...) %>%
+    skimr::yank("numeric")
 }