+++ data-raw/DataComputing.R
@@ -1,18 +1,19 @@
 library(Rd2roxygen)
-library(mosaic)
+library(tidyverse)
 
 # MedicareProviders
 download.file("https://github.com/DataComputing/DataComputing/raw/master/data/MedicareProviders.rda", "data/MedicareProviders.rda")
 cat(create_roxygen(parse_file("https://raw.githubusercontent.com/DataComputing/DataComputing/master/man/MedicareProviders.Rd")), sep = "\n")
 
 # MedicareCharges
-download.file("https://github.com/DataComputing/DataComputing/raw/master/data/MedicareCharges.rda", "inst/extdata/MedicareCharges.RData")
-load("inst/extdata/MedicareCharges.RData")
+download.file("https://github.com/DataComputing/DataComputing/raw/master/data/MedicareCharges.rda", "data-raw/MedicareCharges.RData")
+load("data-raw/MedicareCharges.RData")
 MedicareCharges <- MedicareCharges %>%
-  inner_join(MedicareProviders, by = "idProvider") %>%
+  inner_join(mdsr::MedicareProviders, by = "idProvider") %>%
   group_by(drg, stateProvider) %>%
-  summarize(num_charges = n(), mean_charge = mean(aveCharges, na.rm = TRUE))
-save(MedicareCharges, file = "data/MedicareCharges.rda", compress = "xz")
+  summarize(num_charges = n(), mean_charge = mean(aveCharges, na.rm = TRUE)) %>%
+  ungroup()
+usethis::use_data(MedicareCharges, overwrite = TRUE)
 
 cat(create_roxygen(parse_file("https://raw.githubusercontent.com/DataComputing/DataComputing/master/man/MedicareCharges.Rd")), sep = "\n")
 
@@ -32,8 +33,14 @@ save(Minneapolis2013, file = "data/Minneapolis2013.rda", compress = "xz")
 download.file("https://github.com/DataComputing/DataComputing/raw/master/data/WorldCities.rda", "data/WorldCities.rda")
 cat(create_roxygen(parse_file("https://raw.githubusercontent.com/DataComputing/DataComputing/master/man/WorldCities.Rd")), sep = "\n")
 load("data/WorldCities.rda")
-Encoding(WorldCities$name) <- "latin1"
-WorldCities$name <- iconv(WorldCities$name, from = "latin1", to = "UTF-8")
+showNonASCII(WorldCities$name)
+x <- WorldCities$name
+Encoding(x) <- "latin1"
+y <- iconv(x, from = "latin1", to = "ASCII", sub = "")
+showNonASCII(y)
+# z <- stringi::stri_trans_general(WorldCities$name, "latin-ascii")
+# showNonASCII(z)
+WorldCities$name <- y
 save(WorldCities, file = "data/WorldCities.rda", compress = "xz")
 
 # NCI60 -- 5 MB file so put it in a different place