+++ inst/rstudio/connections/MDSR_MySQL_Server.R
@@ -1,0 +1,1 @@
+db <- mdsr::src_scidb(${0:Schema="airlines"})