+++ NEWS.md
@@ -1,7 +1,31 @@
-# mdsr 0.1.4
+mdsr 0.1.8
+
+* Removed deprecated versions of `dplyr` functions.
+* Added `skim()`
+
+mdsr 0.1.7
+
+* added Connections pane interface to DB Server [#35](https://github.com/beanumber/mdsr/issues/35)
+* removed superfluous extdata [#37](https://github.com/beanumber/mdsr/issues/37)
+* removed broken link in CIA documentation [#38](https://github.com/beanumber/mdsr/issues/38)
+
+mdsr 0.1.6
+
+* improved documentation [#18](https://github.com/beanumber/mdsr/issues/18)
+* `scidb` functions now point to AWS RDS instance
+* Added tests
+* Removed `tidyverse` dependency [#31](https://github.com/beanumber/mdsr/issues/31)
+
+mdsr 0.1.5
+
+* Fixed [#27](https://github.com/beanumber/mdsr/issues/27): non-ASCII characters
+* Fixed [#26](https://github.com/beanumber/mdsr/issues/26): Namespace import from `tidyverse` not imported from
+
+mdsr 0.1.4
 
 * Added CRAN badge after release.
 * Added `etl_NCI60()` function
+* Migrated to `dbplyr`
 
 # mdsr 0.1.3